refactor: extract business logic for 3DS redirections in a class
it paves the way for reusing this logic in other contexts, such as a Web
API one.

refs #787
