+++ etc/di.xml
@@ -1099,6 +1099,8 @@
                 type="Adyen\Payment\Model\AdyenInitiateTerminalApi"/>
     <preference for="Adyen\Payment\Api\AdyenThreeDS2ProcessInterface"
                 type="Adyen\Payment\Model\AdyenThreeDS2Process"/>
+    <preference for="Adyen\Payment\Api\AdyenThreeDSProcessInterface"
+                type="Adyen\Payment\Model\AdyenThreeDSProcess"/>
     <preference for="Adyen\Payment\Api\AdyenOriginKeyInterface"
                 type="Adyen\Payment\Model\AdyenOriginKey"/>
     <preference for="Adyen\Payment\Api\AdyenOrderPaymentStatusInterface"
